+++ CHANGELOG.md
@@ -7,12 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- programs(store): Added `cumulative_inv_cost_factor` as a new global metric for GT, along with an instruction to update it.
+- programs(store): Added `mint_gt_reward` instruction.
+
+## [0.7.1] - 2025-08-15
+
 ### Breaking Changes
+
 - program(chainlink-datastreams): Record last update diff in seconds instead of nanoseconds.
+
+### Added
+
+- sdk(sdk): Implemented off-chain order simulation.
 
 ### Fixed
 
 - sdk(sdk): Ensured pay token account is prepared for wrap native.
+- sdk(chainlink-datastreams): Used `ReportDataV7` to decode version 7 reports.
 
 ### Changes
 
@@ -321,7 +334,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Provided SDK (`gmsol`) and other utility crates.
 - Provided a command-line interface (`gmsol`).
 
-[unreleased]: https://github.com/gmsol-labs/gmx-solana/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/gmsol-labs/gmx-solana/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.7.1
 [0.7.0]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.7.0
 [0.6.0]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.6.0
 [0.5.0]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.5.0